feat: added a new script file for global test utils & added docker availability check
