+++ apps/coordinator/e2e/utils/globalSetup.ts
@@ -1,53 +1,26 @@
 import {FullConfig} from "@playwright/test"
 import {execSync} from "child_process"
 import bitcoinClient from "./bitcoinClient";
+import createTestWallets, {checkDockerAvailability} from "./testFixtures"
+
 
 const globalWalletData ={
     walletNames: [] as string[],
     testWallets: [] as any[]
 }
 
-
 async function globalSetup(_config: FullConfig){
 
    try {
-    execSync("docker compose up -d", {
-        stdio: "inherit",
-        cwd: process.cwd()
-    })
+
+    await checkDockerAvailability();
+    
     const client = bitcoinClient();
     
     await new Promise(resolve => setTimeout(resolve,2000));
     await client?.waitForBitcoinCore();
 
-    const timestamp = Date.now();
-    
-    const walletNames = [
-        `test_wallet_${timestamp}_1`,
-        `test_wallet_${timestamp}_2`,
-        `test_wallet_${timestamp}_3`,
-        `watcher_wallet_${timestamp}`
-    ]
-
-    for (const walletName in walletNames){
-        try {
-            const exist = await client?.walletexists(walletName);
-        if(exist){
-            // Cleaning up the existing wallets
-            await client?.unloadWallet(walletName);
-        }
-        } catch (error) {
-            console.log(`Wallet ${walletName} doesnt exist or already unloaded`)
-        }
-    }
-
-    const wallet1 = await client?.createWallet({wallet_name: walletNames[0]});
-    const wallet2 = await client?.createWallet({wallet_name: walletNames[1]});
-    const wallet3 = await client?.createWallet({wallet_name: walletNames[2]});
-
-    const watcher_wallet = await client?.createWallet({wallet_name: walletNames[3], disable_private_keys: true})
-
-    const testWallets = [wallet1, wallet2,wallet3,watcher_wallet];
+    const {walletNames, testWallets} = await createTestWallets(client!);
 
     globalWalletData.walletNames = walletNames;
     globalWalletData.testWallets = testWallets;
@@ -56,21 +29,24 @@ async function globalSetup(_config: FullConfig){
     process.env.TEST_WALLET_NAMES = JSON.stringify(walletNames)
     process.env.TEST_WALLETS = JSON.stringify(testWallets)
 
-
-    
    } catch (error) {
     console.log("Global setup failed:", error)
-    try {
-        //cleaning up on failures
-        execSync("docker compose down",{
-            stdio: "inherit",
-            cwd: process.cwd()
-        })
-        
-    } catch (clearupError) {
-        console.log("Error while cleaning up",clearupError)
+    
+    // Only attempt Docker cleanup if Docker is actually available
+    if (error instanceof Error && !error.message.includes("Docker is required")) {
+        try {
+            execSync("docker compose down",{
+                stdio: "inherit",
+                cwd: process.cwd()
+            })
+            
+        } catch (clearupError) {
+            console.log("Error while cleaning up",clearupError)
+        }
     }
+    throw error;
    }
 }
+
 
 export default globalSetup
